+++ src/audio/instrument.ts
@@ -1,4 +1,6 @@
 
+export type { InstrumentSamples } from "web-music-score-samples/core";
+
 /**
  * Instrument interface.
  * <p>To create your own instrument just create a class that implement this simple interface.</p>
@@ -36,23 +38,6 @@ export interface Instrument {
      * Stop playback.
      */
     stop(): void;
-}
-
-/**
- * Instruments can also be added using this interface that holds instrument name and audio samples.
- */
-export interface InstrumentSamples {
-    /**
-     * Get instrument name.
-     * @returns - Instrument name.
-     */
-    getName(): string;
-
-    /**
-     * Get samples e.g. { "E2": "base64 string mp3", "E3": "base64 string mp3", etc. }
-     * @returns - Record of audio samples.
-     */
-    getSamples(): Record<string, string>;
 }
 
 /**