+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 ## [6.4.0-pre.8] - Unreleased
 ### Added
 - Implemented octave shift annotation kinds "8va", "8vb" and "loco".
-- InstrumentSamples class. Use this for sampler instruments, Classical Guitar for now.
+- InstrumentSamples class. Use this for sampler instruments like "Classical Guitar" so they do not need to bundle with Tone.js.
+- Several built-in synth instruments.
 
 ### Changed
 - Updated annotation layout with lanes, and stuff.