+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 # CHANGELOG
-## [6.4.0-pre.8] - Unreleased
+## [6.4.0-pre.8] - 2026-02-27
 ### Added
 - Implemented octave shift annotation kinds "8va", "8vb" and "loco".
 - Several built-in synth instruments.